// generated-api/models/PredictMatchRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Nexus Boxing API
 * REST API for authentication, all time ranked boxers, perfect boxer generation, updating category weights, making predictions, managing prediction histories, and weight classes.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { BoxerInput } from './BoxerInput';
import {
    BoxerInputFromJSON,
    BoxerInputFromJSONTyped,
    BoxerInputToJSON,
    BoxerInputToJSONTyped,
} from './BoxerInput';

/**
 * Match prediction request containing two fighters and the weight class
 * @export
 * @interface PredictMatchRequest
 */
export interface PredictMatchRequest {
    /**
     * 
     * @type {number}
     * @memberof PredictMatchRequest
     */
    weightClassId: number;
    /**
     * 
     * @type {BoxerInput}
     * @memberof PredictMatchRequest
     */
    boxerA?: BoxerInput;
    /**
     * 
     * @type {BoxerInput}
     * @memberof PredictMatchRequest
     */
    boxerB?: BoxerInput;
}

/**
 * Check if a given object implements the PredictMatchRequest interface.
 */
export function instanceOfPredictMatchRequest(value: object): value is PredictMatchRequest {
    if (!('weightClassId' in value) || value['weightClassId'] === undefined) return false;
    return true;
}

export function PredictMatchRequestFromJSON(json: any): PredictMatchRequest {
    return PredictMatchRequestFromJSONTyped(json, false);
}

export function PredictMatchRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PredictMatchRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'weightClassId': json['weightClassId'],
        'boxerA': json['boxerA'] == null ? undefined : BoxerInputFromJSON(json['boxerA']),
        'boxerB': json['boxerB'] == null ? undefined : BoxerInputFromJSON(json['boxerB']),
    };
}

export function PredictMatchRequestToJSON(json: any): PredictMatchRequest {
    return PredictMatchRequestToJSONTyped(json, false);
}

export function PredictMatchRequestToJSONTyped(value?: PredictMatchRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'weightClassId': value['weightClassId'],
        'boxerA': BoxerInputToJSON(value['boxerA']),
        'boxerB': BoxerInputToJSON(value['boxerB']),
    };
}