// generated-api/models/WeightClassResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Nexus Boxing API
 * REST API for authentication, all time ranked boxers, perfect boxer generation, updating category weights, making predictions, managing prediction histories, and weight classes.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface WeightClassResponse
 */
export interface WeightClassResponse {
    /**
     * 
     * @type {number}
     * @memberof WeightClassResponse
     */
    weightClassId?: number;
    /**
     * 
     * @type {string}
     * @memberof WeightClassResponse
     */
    className?: string;
    /**
     * 
     * @type {number}
     * @memberof WeightClassResponse
     */
    maxWeightLb?: number;
    /**
     * 
     * @type {number}
     * @memberof WeightClassResponse
     */
    minWeightLb?: number;
}

/**
 * Check if a given object implements the WeightClassResponse interface.
 */
export function instanceOfWeightClassResponse(value: object): value is WeightClassResponse {
    return true;
}

export function WeightClassResponseFromJSON(json: any): WeightClassResponse {
    return WeightClassResponseFromJSONTyped(json, false);
}

export function WeightClassResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): WeightClassResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'weightClassId': json['weightClassId'] == null ? undefined : json['weightClassId'],
        'className': json['className'] == null ? undefined : json['className'],
        'maxWeightLb': json['maxWeightLb'] == null ? undefined : json['maxWeightLb'],
        'minWeightLb': json['minWeightLb'] == null ? undefined : json['minWeightLb'],
    };
}

export function WeightClassResponseToJSON(json: any): WeightClassResponse {
    return WeightClassResponseToJSONTyped(json, false);
}

export function WeightClassResponseToJSONTyped(value?: WeightClassResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'weightClassId': value['weightClassId'],
        'className': value['className'],
        'maxWeightLb': value['maxWeightLb'],
        'minWeightLb': value['minWeightLb'],
    };
}